@@ -25,21 +25,24 @@
     "url": "https://github.com/dantrain/react-stonecutter.git"
   },
   "devDependencies": {
-    "babel-core": "^6.11.4",
+    "babel-core": "^6.13.2",
+    "babel-eslint": "^6.1.2",
     "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-es2015": "^6.13.2",
     "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-2": "^6.11.0",
+    "babel-preset-stage-2": "^6.13.0",
     "browser-sync": "^2.14.0",
     "chai": "^3.5.0",
     "chai-enzyme": "^0.5.0",
     "enzyme": "^2.4.1",
-    "eslint": "^2.13.1",
-    "eslint-config-airbnb": "^6.2.0",
-    "eslint-plugin-react": "^4.2.3",
+    "eslint": "^3.2.2",
+    "eslint-config-airbnb": "^10.0.0",
+    "eslint-plugin-import": "^1.12.0",
+    "eslint-plugin-jsx-a11y": "^2.0.1",
+    "eslint-plugin-react": "^6.0.0",
     "ghooks": "^1.3.2",
     "gulp": "^3.9.1",
-    "gulp-eslint": "^2.0.0",
+    "gulp-eslint": "^3.0.1",
     "gulp-filter": "^4.0.0",
     "gulp-rename": "^1.2.2",
     "gulp-shell": "^0.5.2",
@@ -48,12 +51,10 @@
     "lodash.camelcase": "^4.2.0",
     "lodash.isequalwith": "^4.3.1",
     "lodash.shuffle": "^4.1.0",
-    "mocha": "^2.5.3",
-    "node-libs-browser": "^1.0.0",
+    "mocha": "^3.0.2",
     "node-notifier": "^4.6.0",
     "rc-slider": "^3.8.2",
     "react": "^15.3.0",
-    "react-addons-test-utils": "^15.3.0",
     "react-dom": "^15.3.0",
     "react-radio-group": "^3.0.1",
     "run-sequence": "^1.2.2",
