@@ -1,20 +1,23 @@
-import React from 'react';
+import React, { Component } from 'react';
+
 const enquire = typeof window !== 'undefined' ? require('enquire.js') : null;
 
 export default (Grid, { maxWidth, minPadding = 0, defaultColumns = 4 } = {}) =>
-  React.createClass({
+  class extends Component {
 
-    getInitialState() {
-      return {
+    constructor(props) {
+      super(props);
+
+      this.state = {
         columns: defaultColumns
       };
-    },
+    }
 
     componentDidMount() {
       const { columnWidth, gutterWidth } = this.props;
 
       const breakpoints = [];
-      const getWidth = i => i * (columnWidth + gutterWidth) - gutterWidth + minPadding;
+      const getWidth = i => (i * (columnWidth + gutterWidth)) - gutterWidth + minPadding;
 
       for (let i = 1; getWidth(i) <= maxWidth + columnWidth + gutterWidth; i++) {
         breakpoints.push(getWidth(i));
@@ -33,15 +36,15 @@ export default (Grid, { maxWidth, minPadding = 0, defaultColumns = 4 } = {}) =>
 
       this.breakpoints.forEach(({ breakpoint, handler }) =>
         enquire.register(breakpoint, { match: handler }));
-    },
+    }
 
     componentWillUnmount() {
       this.breakpoints.forEach(({ breakpoint, handler }) =>
         enquire.unregister(breakpoint, handler));
-    },
+    }
 
     render() {
       return <Grid {...this.props} {...this.state} />;
     }
 
-  });
+  };
