@@ -14,7 +14,7 @@ export default function(items, props) {
       throw new Error('Each child must have an "itemHeight" prop or an "itemRect.height" prop.');
     }
 
-    const x = column * columnWidth + column * gutterWidth;
+    const x = (column * columnWidth) + (column * gutterWidth);
     const y = columnHeights[column];
 
     columnHeights[column] += Math.round(height) + gutterHeight;
@@ -22,7 +22,7 @@ export default function(items, props) {
     return [x, y];
   });
 
-  const gridWidth = columns * columnWidth + ((columns - 1) * gutterWidth);
+  const gridWidth = (columns * columnWidth) + ((columns - 1) * gutterWidth);
   const gridHeight = Math.max.apply(null, columnHeights) - gutterHeight;
 
   return { positions, gridWidth, gridHeight };
